Began implementation of navigation for instructions
