+++ BuildEnvironment/app/src/main/java/com/example/h2project/HomeScreenFragment1.java
@@ -24,11 +24,11 @@ public class HomeScreenFragment1 extends Fragment {
     public void onViewCreated(@NonNull View view, Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
 
-        view.findViewById(R.id.home_screen_fragment_1_button_1).setOnClickListener(new View.OnClickListener() {
+        view.findViewById(R.id.home_screen_fragment_1_button_4).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Intent intent = new Intent(getActivity(), ScavengerHunt.class);
-                startActivity(intent);
+                NavHostFragment.findNavController(HomeScreenFragment1.this)
+                        .navigate(R.id.action_homeScreenFragment1_to_instructionsFragment1);
             }
         });
     }
