+++ BuildEnvironment/app/src/main/res/layout/instructions_fragment_1.xml
@@ -11,4 +11,10 @@
         android:layout_height="match_parent"
         android:text="@string/hello_blank_fragment" />
 
+    <Button
+        android:id="@+id/instructions_fragment_1_button_1"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Button1" />
+
 </FrameLayout>
