+++ BuildEnvironment/app/src/main/res/navigation/nav_graph.xml
@@ -3,86 +3,51 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/nav_graph"
-    app:startDestination="@id/FirstFragment">
+    app:startDestination="@id/homeScreenFragment1">
 
     <fragment
-        android:id="@+id/FirstFragment"
+        android:id="@+id/instructionsFragment1"
+        android:name="com.example.h2project.InstructionsFragment1"
+        android:label="instructions_fragment_1"
+        tools:layout="@layout/instructions_fragment_1" >
+        <action
+            android:id="@+id/action_instructionsFragment1_to_instructionsFragment2"
+            app:destination="@id/instructionsFragment2" />
+    </fragment>
+    <fragment
+        android:id="@+id/instructionsFragment2"
+        android:name="com.example.h2project.InstructionsFragment2"
+        android:label="instructions_fragment_2"
+        tools:layout="@layout/instructions_fragment_2" >
+        <action
+            android:id="@+id/action_instructionsFragment2_to_instructionsFragment3"
+            app:destination="@id/instructionsFragment3" />
+    </fragment>
+    <fragment
+        android:id="@+id/instructionsFragment3"
+        android:name="com.example.h2project.InstructionsFragment3"
+        android:label="instructions_fragment_3"
+        tools:layout="@layout/instructions_fragment_3" >
+        <action
+            android:id="@+id/action_instructionsFragment3_to_instructionsFragment4"
+            app:destination="@id/instructionsFragment4" />
+    </fragment>
+    <fragment
+        android:id="@+id/instructionsFragment4"
+        android:name="com.example.h2project.InstructionsFragment4"
+        android:label="instructions_fragment_4"
+        tools:layout="@layout/instructions_fragment_4" >
+        <action
+            android:id="@+id/action_instructionsFragment4_to_homeScreenFragment1"
+            app:destination="@id/homeScreenFragment1" />
+    </fragment>
+    <fragment
+        android:id="@+id/homeScreenFragment1"
         android:name="com.example.h2project.HomeScreenFragment1"
-        android:label="@string/first_fragment_label"
-        tools:layout="@layout/HomeScreenFragment1">
-
+        android:label="home_screen_fragment_1"
+        tools:layout="@layout/home_screen_fragment_1" >
         <action
-            android:id="@+id/action_FirstFragment_to_SecondFragment"
-            app:destination="@id/SecondFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/SecondFragment"
-        android:name="com.example.h2project.HomeScreenFragment2"
-        android:label="@string/second_fragment_label"
-        tools:layout="@layout/home_screen_fragment_2">
-
-        <action
-            android:id="@+id/action_SecondFragment_to_FirstFragment"
-            app:destination="@id/FirstFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/FirstFragment"
-        android:name="com.example.h2project.ScavengerHuntFragment1"
-        android:label="@string/first_fragment_label"
-        tools:layout="@layout/scavenger_hunt_fragment_1">
-
-        <action
-            android:id="@+id/action_FirstFragment_to_SecondFragment"
-            app:destination="@id/SecondFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/SecondFragment"
-        android:name="com.example.h2project.ScavengerHuntFragment2"
-        android:label="@string/second_fragment_label"
-        tools:layout="@layout/scavenger_hunt_fragment_2">
-
-        <action
-            android:id="@+id/action_SecondFragment_to_FirstFragment"
-            app:destination="@id/FirstFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/FirstFragment"
-        android:name="com.example.h2project.FirstFragment"
-        android:label="@string/first_fragment_label"
-        tools:layout="@layout/fragment_first">
-
-        <action
-            android:id="@+id/action_FirstFragment_to_SecondFragment"
-            app:destination="@id/SecondFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/SecondFragment"
-        android:name="com.example.h2project.SecondFragment"
-        android:label="@string/second_fragment_label"
-        tools:layout="@layout/fragment_second">
-
-        <action
-            android:id="@+id/action_SecondFragment_to_FirstFragment"
-            app:destination="@id/FirstFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/First2Fragment"
-        android:name="com.example.h2project.First2Fragment"
-        android:label="@string/first_fragment_label"
-        tools:layout="@layout/fragment_first2">
-
-        <action
-            android:id="@+id/action_First2Fragment_to_Second2Fragment"
-            app:destination="@id/Second2Fragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/Second2Fragment"
-        android:name="com.example.h2project.Second2Fragment"
-        android:label="@string/second_fragment_label"
-        tools:layout="@layout/fragment_second2">
-
-        <action
-            android:id="@+id/action_Second2Fragment_to_First2Fragment"
-            app:destination="@id/First2Fragment" />
+            android:id="@+id/action_homeScreenFragment1_to_instructionsFragment1"
+            app:destination="@id/instructionsFragment1" />
     </fragment>
 </navigation>
